feat: allow separate usage of general command without importing dates
